✅ 이벤트 발송 테스트 추가
